Async API is intoduced for Route4MeManager and Route4MeManagerV5.
Mix of sync/async context is removed for async flow.
IDs are switched from 32 bit integer to 64 bit integer (both signed and unsigned).
Async API for FastProcessing (JSON) is introduced.
Unnesessary internal state is removed from FastProcessing.
ArrayList is removed from geocoding API.
Internal sleeps and spin waits are removed.
Naming is updated to follow CamelCase.
Internal code smells are removed.
NewtonJson usage is fixed.

+++ route4me-csharp-sdk/Route4MeSDKLibrary/DataTypes/OrderCustomField.cs
@@ -13,7 +13,7 @@ namespace Route4MeSDK.DataTypes
         ///     Custom order field ID.
         /// </summary>
         [DataMember(Name = "order_custom_field_id", EmitDefaultValue = false)]
-        public int OrderCustomFieldId { get; set; }
+        public long OrderCustomFieldId { get; set; }
 
         /// <summary>
         ///     Custom order field name.
